finance-data-platform/phase_6_streaming_kafka/requirements.txt
kafka-python==2.0.2
boto3>=1.42.55
python-dotenv==1.0.0
faker>=20.0.0
pandas>=1.5.0
openpyxl>=3.1.0
